16-sling-data/pyproject.toml
[project]
name = "16-sling-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sling>=1.4.26",
    "tower>=0.3.35",
]
